#310 Moved properties replacement into PropertiesReplacer.
